refactor: Clean code
- merge `ModuleData` into `PackageData`
- add doc comments to public interfaces
- fix license headers
- make it pass clippy
