+++ crates/sui-genesis-builder/src/stardust/error.rs
@@ -1,4 +1,3 @@
-// Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 //! Error types pertaining to deserializing Stardust snapshots
