+++ crates/sui-genesis-builder/src/stardust/native_token/mod.rs
@@ -1,8 +1,7 @@
-// Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 //! The [`native_token`] module incorporates all the logic necessary for
 //! building Stardust native token modules and packages.
-pub mod module_data;
+
 pub mod package_builder;
 pub mod package_data;
