+++ crates/sui-genesis-builder/src/stardust/native_token/package_builder.rs
@@ -12,18 +12,17 @@ use sui_move_build::{BuildConfig, CompiledPackage};
 
 use crate::stardust::native_token::package_data::NativeTokenPackageData;
 
+/// The [`PackageBuilder`] struct is responsible for building and compiling Stardust native token packages.
 pub struct PackageBuilder;
 impl PackageBuilder {
-    pub fn new() -> Self {
-        Self
-    }
 
+    /// Builds and compiles a Stardust native token package.
     pub fn build_and_compile(&self, package: NativeTokenPackageData) -> Result<CompiledPackage> {
         let package_template_path =
             Path::new("crates/sui-genesis-builder/src/stardust/native_token/package_template");
 
         let new_package_path = Path::new("crates/sui-genesis-builder/src/stardust/native_token")
-            .join(&format!(
+            .join(format!(
                 "native_token_package_{}",
                 package.module().native_token_id()
             ));
@@ -43,6 +42,7 @@ impl PackageBuilder {
         Ok(compiled_package)
     }
 
+    // Copies the contents of a directory to a new location.
     fn copy_template_dir(from: &Path, to: &PathBuf) -> Result<()> {
         // Step 1: Create a new folder for the package and copy the contents
         if to.exists() {
@@ -50,9 +50,9 @@ impl PackageBuilder {
                 "Path {} already exists. Deleting the existing directory.",
                 to.display()
             );
-            fs::remove_dir_all(&to)?;
+            fs::remove_dir_all(to)?;
         }
-        fs::create_dir_all(&to.join("sources"))?;
+        fs::create_dir_all(to.join("sources"))?;
 
         // Recursive copy function to handle directories and files
         fn recursive_copy(src: &Path, dst: &Path) -> io::Result<()> {
@@ -72,32 +72,34 @@ impl PackageBuilder {
             }
             Ok(())
         }
-        recursive_copy(&from, &to)?;
+        recursive_copy(from, to)?;
 
         Ok(())
     }
 
-    fn adjust_move_toml(package_path: &PathBuf, package: &NativeTokenPackageData) -> Result<()> {
+    // Adjusts the Move.toml file with the package name and alias address.
+    fn adjust_move_toml(package_path: &Path, package: &NativeTokenPackageData) -> Result<()> {
         let cargo_toml_path = package_path.join("Move.toml");
         let contents = fs::read_to_string(&cargo_toml_path)?;
         let new_contents = contents
-            .replace("$PACKAGE_NAME", &package.move_toml().package_name())
-            .replace("$ALIAS", &package.module().alias_address());
+            .replace("$PACKAGE_NAME", package.move_toml().package_name())
+            .replace("$ALIAS", package.module().alias_address());
         fs::write(&cargo_toml_path, new_contents)?;
 
         Ok(())
     }
 
+    // Replaces template variables in the .move file with the actual values.
     fn adjust_native_token_module(
-        package_path: &PathBuf,
+        package_path: &Path,
         package: &NativeTokenPackageData,
     ) -> Result<()> {
         let old_move_file_path = package_path.join("sources/native_token_template.move");
         let new_move_file_name = format!("{}.move", package.module().module_name());
-        let new_move_file_path = package_path.join("sources").join(&new_move_file_name);
+        let new_move_file_path = package_path.join("sources").join(new_move_file_name);
 
         // Rename the template .move file
-        fs::rename(&old_move_file_path, &new_move_file_path)?;
+        fs::rename(old_move_file_path, &new_move_file_path)?;
 
         let contents = fs::read_to_string(&new_move_file_path)?;
 
@@ -110,14 +112,14 @@ impl PackageBuilder {
         };
 
         let new_contents = contents
-            .replace("$MODULE_NAME", &package.module().module_name())
-            .replace("$OTW", &package.module().otw_name())
+            .replace("$MODULE_NAME", package.module().module_name())
+            .replace("$OTW", package.module().otw_name())
             .replace("$COIN_DECIMALS", &package.module().decimals().to_string())
-            .replace("$COIN_SYMBOL", &package.module().symbol())
-            .replace("$COIN_NAME", &package.module().coin_name())
-            .replace("$COIN_DESCRIPTION", &package.module().coin_description())
+            .replace("$COIN_SYMBOL", package.module().symbol())
+            .replace("$COIN_NAME", package.module().coin_name())
+            .replace("$COIN_DESCRIPTION", package.module().coin_description())
             .replace("$ICON_URL", &icon_url)
-            .replace("$ALIAS_ADDRESS", &package.module().alias_address());
+            .replace("$ALIAS_ADDRESS", package.module().alias_address());
 
         fs::write(&new_move_file_path, new_contents)?;
 
@@ -145,7 +147,7 @@ mod tests {
         let src_dir = test_package_path.join("sources");
         fs::create_dir_all(&src_dir).unwrap();
         let test_file_path = src_dir.join("test.move");
-        let mut file = File::create(&test_file_path).unwrap();
+        let mut file = File::create(test_file_path).unwrap();
         writeln!(file, "0x0::test {{}}").unwrap();
 
         // Define the target directory for the files to be copied
