+++ crates/sui-genesis-builder/src/stardust/parse.rs
@@ -1,4 +1,3 @@
-// Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 //! Types and logic to parse a full Stardust snapshot.
