+++ crates/sui-genesis-builder/src/stardust/types.rs
@@ -1,4 +1,3 @@
-// Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 //! Types representing blocks of data in a Stardust snapshot.
